fix: fix controllers validations

+++ backend/src/controllers/DELETE/product.ts
@@ -1,5 +1,7 @@
+import { Product } from '../../models/product'
 import { CreateLogRepository } from '../../repositories/CREATE/log'
 import { RemoveProductRepository } from '../../repositories/REMOVE/product'
+import { SelectProductRepository } from '../../repositories/SELECT/product'
 import { getIDFromToken } from '../../tools/getUserFromToken'
 import { badRequest, headersAuthError, internalError, successfull } from '../helpers'
 import { type HTTPRequest, type HTTPResponse } from '../protocols'
@@ -13,6 +15,10 @@ export class DeleteProductController {
         if (httpRequest.params === undefined) {
           return badRequest('No parameters was provided')
         } else {
+          const product: Product = await new SelectProductRepository().selectOne(+httpRequest.params.id)
+          if (product === undefined || product === null) {
+            return badRequest('Product not found')
+          }
           await new RemoveProductRepository().remove(+httpRequest.params.id)
           const id = await getIDFromToken(httpRequest.headers.authorization)
           await new CreateLogRepository().create({
